+++ src/main/java/com/elmakers/mine/bukkit/spell/builtin/DisarmSpell.java
@@ -22,6 +22,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 
+@Deprecated
 public class DisarmSpell extends TargetingSpell 
 {
     private Random random = new Random();